refactor qlr generation function
